backend/requirements.txt
fastapi==0.115.*
uvicorn==0.34.*
python-dotenv==1.1.*
openai==1.82.*
python-multipart==0.0.*
requests==2.32.*
pydantic==2.11.*
Pillow==11.2.*
pytesseract==0.3.*
numpy==2.2.*
scipy==1.15.*
